Refactored extensions and added StringUtils for building UUIDs
